reorg

+++ dotfiles/.zshrc
@@ -1,25 +1,28 @@
 source $HOME/.commonrc
 
-ZSH=$HOME/.oh-my-zsh
-ZSH_THEME="nicoulaj"
 CASE_SENSITIVE="true"
-
 HISTSIZE=1000
 SAVEHIST=1000
 
-setopt APPEND_HISTORY
 setopt INC_APPEND_HISTORY
 setopt SHARE_HISTORY
 setopt NO_BEEP
 setopt EXTENDED_GLOB
 setopt HIST_IGNORE_DUPS
 setopt NONOMATCH
+
 autoload -U zmv
 
-plugins=(git brew cp django extract fabric git-flow npm osx pep8 pip python dirhistory gem tmuxinator vagrant)
+# oh my zsh
+OHMYZSH=$HOME/.oh-my-zsh/oh-my-zsh.sh
+if [ -f $OHMYZSH ]; then
+    source $OHMYZSH
 
-source $ZSH/oh-my-zsh.sh
+    ZSH_THEME="nicoulaj"
+    plugins=(git brew cp django extract fabric git-flow npm osx pep8 pip python dirhistory gem tmuxinator vagrant)
+fi
 
+# zsh pip completion
 function _pip_completion {
   local words cword
   read -Ac words
